+++ packages/cli/src/commands/service/dev.ts
@@ -14,7 +14,7 @@ import version from '../../version'
 
 const ipfsClient = require('ipfs-http-client')
 
-type Context = Environment.IStop | Service.ICompile | Service.ICreate | Runner.ICreate | Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs
+type Context = Environment.IStart | Service.ICompile | Service.ICreate | Runner.ICreate | Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs
 type ContextEnd = Runner.ILogsStop | Instance.IEventLogsStop | Runner.IResultLogsStop | Environment.IStop
 
 export default class Dev extends Command {