+++ packages/cli/CHANGELOG.md
@@ -41,6 +41,8 @@ Commands `service:dev` and `process:dev` are automatically start a local environ
 
 #### Bug fixes
 
+- [#191](https://github.com/mesg-foundation/js-sdk/pull/191) Do not stop the environment if other instances of the CLI rely on it
+
 ## [v0.3.0](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%400.3.0)
 
 #### Improvements